Выносит классы из api.service.js в отдельные файлы.
