+++ src/frontend/src/services/api.service.js
@@ -1,65 +1,4 @@
-import JwtService from "@/services/jwt.service";
-import axios from "@/plugins/axios";
+import { AuthApiService } from "@/services/api_classes/auth-api-service";
+import { CrudApiService } from "@/services/api_classes/crud-api-service";
 
-class BaseApiService {
-  constructor(notifier) {
-    if (!axios.$notifier) {
-      axios.$notifier = notifier;
-    }
-  }
-}
-
-export class AuthApiService extends BaseApiService {
-  constructor(notifier) {
-    super(notifier);
-  }
-  setAuthHeader() {
-    const token = JwtService.getToken();
-    axios.defaults.headers.common["Authorization"] = token
-      ? `Bearer ${token}`
-      : ``;
-  }
-
-  async login(params) {
-    const { data } = await axios.post("login", params);
-    return data;
-  }
-
-  async logout() {
-    const { data } = await axios.delete("logout");
-    return data;
-  }
-
-  async getMe() {
-    const { data } = await axios.get("whoAmI");
-    return data;
-  }
-}
-
-export class CrudApiService extends BaseApiService {
-  #resource;
-  constructor(resource, notifier) {
-    super(resource, notifier);
-    this.#resource = resource;
-  }
-
-  async get() {
-    const { data } = await axios.get(this.#resource);
-    return data;
-  }
-
-  async post(entity) {
-    const { data } = await axios.post(this.#resource, entity);
-    return data;
-  }
-
-  async put(entity) {
-    const { data } = await axios.put(`${this.#resource}/${entity.id}`, entity);
-    return data;
-  }
-
-  async delete(id) {
-    const { data } = await axios.delete(`${this.#resource}/${id}`);
-    return data;
-  }
-}
+export { AuthApiService, CrudApiService };
