+++ src/frontend/src/services/api_classes/auth-api-service.js
@@ -1,0 +1,30 @@
+import JwtService from "@/services/jwt.service";
+import axios from "@/plugins/axios";
+import { BaseApiService } from "@/services/api_classes/base-api-service";
+
+export class AuthApiService extends BaseApiService {
+  constructor(notifier) {
+    super(notifier);
+  }
+  setAuthHeader() {
+    const token = JwtService.getToken();
+    axios.defaults.headers.common["Authorization"] = token
+      ? `Bearer ${token}`
+      : ``;
+  }
+
+  async login(params) {
+    const { data } = await axios.post("login", params);
+    return data;
+  }
+
+  async logout() {
+    const { data } = await axios.delete("logout");
+    return data;
+  }
+
+  async getMe() {
+    const { data } = await axios.get("whoAmI");
+    return data;
+  }
+}
