+++ src/frontend/src/services/api_classes/base-api-service.js
@@ -1,0 +1,9 @@
+import axios from "@/plugins/axios";
+
+export class BaseApiService {
+  constructor(notifier) {
+    if (!axios.$notifier) {
+      axios.$notifier = notifier;
+    }
+  }
+}
