+++ src/frontend/src/services/api_classes/crud-api-service.js
@@ -1,0 +1,30 @@
+import { BaseApiService } from "@/services/api_classes/base-api-service";
+import axios from "@/plugins/axios";
+
+export class CrudApiService extends BaseApiService {
+  #resource;
+  constructor(resource, notifier) {
+    super(resource, notifier);
+    this.#resource = resource;
+  }
+
+  async get() {
+    const { data } = await axios.get(this.#resource);
+    return data;
+  }
+
+  async post(entity) {
+    const { data } = await axios.post(this.#resource, entity);
+    return data;
+  }
+
+  async put(entity) {
+    const { data } = await axios.put(`${this.#resource}/${entity.id}`, entity);
+    return data;
+  }
+
+  async delete(id) {
+    const { data } = await axios.delete(`${this.#resource}/${id}`);
+    return data;
+  }
+}
